QENG-4823) Combine scripts and improve docs
This commit combines the scripts into a single resource to limit
duplication; this means that the default lambda_handler definition must
change to `enforce` and `terminate`. Also included in this PR is a
cleanup of the documentation, where there is a cleaner separation for
docs for AWS users and docs for people implementing the Reaper.