(QENG-4823) Fix date validation
There was an error with the `validate_ec2_termination_date` method
previous to this commit; it asked for a `termination_date` string in the
code, but the docs actually specified an ec2 instance. Now it
consistently takes an ec2 instance, as it will call the
`terminate_instance` method if it cannot parse the date.