(QENG-5244) Add cloudformation templates
This commit adds cloudformation templates for reaper deployment; prior
to this, deployment had to be manually done. Now the reaper can be
deployed from a single AWS account into multiple accounts and multiple
regions.

Besides changes to the README, the actual lambda code had to be adjusted
to determine which region it was reporting from, and changing the
environment variables per lambda because cloudformation does not allow
for the underscore character for environment variables.